requirements.txt
asgiref==3.11.1
certifi==2026.1.4
charset-normalizer==3.4.4
cloudinary==1.44.1
crispy-bootstrap5==2025.6
dj-database-url==3.1.0
dj3-cloudinary-storage==0.0.6
Django==5.2.11
django-allauth==65.14.3
django-crispy-forms==2.5
gunicorn==25.1.0
idna==3.11
packaging==26.0
psycopg2-binary==2.9.11
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
urllib3==2.6.3
whitenoise==6.11.0
